manager: track full details of logout request
When we get a logout request, there are two bits of information
that come with the request:

1) Whether or not we should seek confirmation by default from the user
before proceeding
2) Whether or not individual applications should be given an
opportunity to ask us to seek confirmation from the user before
proceeding.

We track the latter in a boolean state variable (forceful_logout),
but not the former.

We already have an enumeration type (GsmManagerLogoutMode) that can
be used to track both bits, though.

This commit changes the boolean state variable to the enum type, so
the whether-we-should-confirmation information is available deeper
down in the logout process.

Based on gnome-session commit: d7adaa9111d0f42d11225c7e29a7ee1d37d74d2f
From: Ray Strode <rstrode@redhat.com>