Restore classic GTK+ 2 style to Log Out and Shut Down dialogs
In days gone by, the MATE Session Log Out and Shut Down dialog boxes
looked similar to other GTK dialogs of the era:  They contained an
icon on the far left, conservatively-sized buttons on the bottom, and
the progress text was drawn over the progress bar.  Then when GTK+ 3
came along, though the coding of the dialog did not change, GTK+ 3
enforced a new style upon the dialog:  No icon, oversized buttons along
the bottom of the dialog, and text placed above the progress bar.  This
commit restores the old-fashioned GTK+ 2 appearance of the dialog, even
on GTK+ 3.