GsmLogoutDialog: Ensure progress bar is tall enough to contain label text
This is a massive kludge.  Whenever the `size-allocate` signal is
emitted by our custom progress bar label, find the height of the
label and then use the `min-height` CSS property on the progress bar
to ensure the progress trough and indicator are tall enough to fit
the text.  Otherwise, ~20-point text tends to "fall off" the progress
bar into the gray area of the window and that looks ugly.